Pull request #278: Fix instrumented test
Merge in ADGUARD-CORE-LIBS/vpn-libs from fix/instrumented_test_problems to master

Squashed commit of the following:

commit aaf60952d27d50f1716f0a7583b0f9ac71d92f26
Author: Max Grupper <77971315+grumaxxx@users.noreply.github.com>
Date:   Fri Feb 2 14:49:43 2024 +0200

    skipci: fix comment

commit 6ae84f6b574806f88322691f2b58aae87788ff04
Author: Max Grupper <77971315+grumaxxx@users.noreply.github.com>
Date:   Fri Feb 2 12:56:23 2024 +0200

    skipci: fix

commit 68a76ad0b017797853b261115587297a398add78
Author: Max Grupper <77971315+grumaxxx@users.noreply.github.com>
Date:   Fri Feb 2 12:35:43 2024 +0200

    skipci: fix

commit baef4527534955c8d71126d5a5b7f4691799f76e
Author: Max Grupper <77971315+grumaxxx@users.noreply.github.com>
Date:   Fri Feb 2 12:06:02 2024 +0200

    skipci: try with relay from answer

commit 5bf393c19f905d8b9cb56bffd14c789aed57d9e1
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 20:24:06 2024 +0200

    skipci: wip

commit 64cbc80406c39f856bcdf3d2b647a8321f0dd3e5
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 19:53:33 2024 +0200

    skipci: increase timeout

commit 7a607c76acea698c5f7188e4e4eb2ee054e50475
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 16:40:33 2024 +0200

    skipci: rkn

commit 0d13fa15a56c4fce5d5c09e3e231f36632819380
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 16:19:39 2024 +0200

    skipci: wip

commit de0a10615538002bc7f22a31afd5fea763a3f10e
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 15:41:06 2024 +0200

    skipci: wip

commit 2b54922ecfa21af500da741d36c29505aa118e91
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 15:18:32 2024 +0200

    skipci: wip

commit abf532b614d36cdfb5209cbc778b0d019b3a7de1
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 15:03:34 2024 +0200

    skipci: wip

commit 67e5959e7983eb653372695d675558050b2a59f5
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 13:57:08 2024 +0200

    skipci: wip

commit 669f01c45bdd06d497586a1e92177f8b2a64148b
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 13:37:51 2024 +0200

    skipci: wip

commit d6d3cc2109506517b37e34e32794052d664cd0a6
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 13:19:38 2024 +0200

    skipci: wip

commit 51a7eaa2d07b6a75fdba2d89f8e955aeed27e7d5
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 12:49:00 2024 +0200

    skipci: wip

commit ef183f20c1bcd83a0383c3cb48a41a25ef02e246
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 12:40:09 2024 +0200

    skipci: wip

commit 26838a2297a8e3ba78a0187e21a33507bc125555
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 12:24:32 2024 +0200

    skipci: rollback

commit e66f3f2ec592d18feaa7009673022468b902379d
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 12:06:50 2024 +0200

    skipci: bamboo

commit e5869be29b86a72626a61b85fd8325e378a0f572
Author: Maxim Grupper <m.grupper@adguard.com>
Date:   Thu Feb 1 12:02:59 2024 +0200

    skipci: add flags to get_location script

+++ integration-tests/integration_test.sh
@@ -204,7 +204,7 @@ get_location_data() {
     docker exec ${COMMON_CONTAINER} /bin/bash -c "bash ./get_location.sh ${ADGUARD_API_LOCATIONS_URL}"
     docker cp ${COMMON_CONTAINER}:result.txt network-load-simulator/endpoint.txt
     ENDPOINT_HOSTNAME=$(cat network-load-simulator/endpoint.txt | grep ENDPOINT_HOSTNAME | cut -d '=' -f 2)
-    ENDPOINT_IP=$BAMBOO_ADGUARD_RELAY_IP
+    ENDPOINT_IP=$(cat network-load-simulator/endpoint.txt | grep ENDPOINT_IP | cut -d '=' -f 2)
     docker stop ${COMMON_CONTAINER}
     rm -f network-load-simulator/endpoint.txt
 }
@@ -216,6 +216,7 @@ run_browser_test() {
   ADGUARD_API_DOMAIN=$BAMBOO_ADGUARD_API_DOMAIN
   ADGUARD_API_CREDS_URL="https://${ADGUARD_API_DOMAIN}${BAMBOO_ADGUARD_API_CREDS_PATH}"
   ADGUARD_API_LOCATIONS_URL="https://${ADGUARD_API_DOMAIN}${BAMBOO_ADGUARD_API_LOCATIONS_PATH}"
+  # Get location data from backend. Hostname and relay IP address of the endpoint
   get_location_data
 
   PROTOCOL=http2
@@ -228,10 +229,22 @@ run_browser_test() {
   run_client_with_browser
 
   sleep 5
+
+  # Check that client is running
+  if ! docker exec "$CLIENT_WITH_BROWSER_CONTAINER" pgrep standalone > /dev/null;
+  then
+    echo "Client is not running"
+    # Try to get some logs from client
+    docker logs $CLIENT_WITH_BROWSER_CONTAINER
+    exit 1
+  fi
+
+  # Run tests for 30 minutes
   docker exec -w /test -e TIME_LIMIT=30m "$CLIENT_WITH_BROWSER_CONTAINER" node index.js || RESULT=1
   docker exec "$CLIENT_WITH_BROWSER_CONTAINER" chmod -R 777 /output/
   docker cp "$CLIENT_WITH_BROWSER_CONTAINER":/test/output.json ./output1part.json
 
+  # Imitate network problems. Drop all traffic to endpoint. Client should reconnect.
   docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c "iptables -A OUTPUT -j DROP; iptables -A INPUT -j DROP"
   sleep 1
   docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c 'pids=$(pgrep standalone); echo "PIDS: $pids"; for pid in $pids; do kill -SIGHUP $pid || true; done'
@@ -239,6 +252,7 @@ run_browser_test() {
   docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c "iptables -D OUTPUT -j DROP; iptables -D INPUT -j DROP"
   sleep 60
 
+  # Run tests again
   docker exec -w /test -e TIME_LIMIT=30m "$CLIENT_WITH_BROWSER_CONTAINER" node index.js || RESULT=1
   docker cp $CLIENT_WITH_BROWSER_CONTAINER:/test/output.json ./output2part.json
   docker stop "$CLIENT_WITH_BROWSER_CONTAINER"
